pyproject.toml
[project]
name = "pgweather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "geopandas>=0.10.2",
    "netcdf4>=1.7.2",
    "numpy>=2.0.2",
    "pandas>=1.1.3",
    "plotly>=4.12.0",
    "python-dateutil>=2.8.1",
    "requests>=2.24.0",
    "scipy>=1.13.1",
    "shapely>=2.0.7",
    "siphon>=0.9",
    "xarray>=2024.7.0",
    "polars>=1.24.0",
    "dotenv>=0.9.9",
    "adbc-driver-manager>=1.5.0",
    "adbc-driver-postgresql>=1.5.0",
    "pyarrow>=18.0.0",
    "geojson>=3.2.0",
    "pytz>=2025.1",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "dash>=3.2.0",
    "pydantic-settings>=2.11.0",
    "cfgrib>=0.9.15.1",
    "eccodes>=2.45.0",
]
